pyproject.toml
[tool.poetry]
name = "siy"
version = "0.1.0"
description = "A system for quickly creating data pipelines in the cloud"
authors = ["Mathieson Sterling <mathieson.sterling@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
apache-airflow = "^2.2.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
